+++ tests/samples/project/exec.py
@@ -1,0 +1,4 @@
+from machinable import Execution
+
+class Test(Execution):
+    pass